LayoutTests/fast/security-labels/window-postmessage-expected.txt

This tests that label propagation when sending message to windows.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


checkpostmessage
checkpostmessage
PASS window.postMessage(message, "*") threw exception Error: SECURITY_ERR: DOM Exception 18.
checkpostmessage
checkpostmessage
PASS window.postMessage(message, "*") threw exception Error: SECURITY_ERR: DOM Exception 18.
checkpostmessage
checkpostmessage
PASS: blocked labeled data
PASS: blocked labeled data
PASS: received ping
PASS successfullyParsed is true

TEST COMPLETE
